+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2024-09-16
+
+### Added
+
+- Application: delete image keyboard support
+- Menu: add "Image" sub-menu
+
+### Fixed
+
+- Footer: minimal index value presentation
+
 ## [0.6.1] - 2024-09-14
 
 ### Changed
@@ -80,7 +91,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Select working directory
 - Show analysis: files type and name formats
 
-[Unreleased]: https://github.com/vikian050194/tart/compare/v0.6.1...HEAD
+[Unreleased]: https://github.com/vikian050194/tart/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/vikian050194/tart/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/vikian050194/tart/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/vikian050194/tart/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/vikian050194/tart/compare/v0.5.0...v0.5.1